separate non-konva components
