+++ src/components/App.js
@@ -1,5 +1,5 @@
 import React from "react";
-import Header from './Header';
+import Header from './header/Header';
 import Canvas from './konva_components/Canvas';
 
 const App = () => {
