+++ src/components/algo_params/VisualizeButton.js
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from 'react';
 import { connect } from "react-redux";
 import { Button } from 'semantic-ui-react';
-import algorithmMappings from '../algorithmMappings';
+import algorithmMappings from '../../algorithmMappings';
 
 const VisualizeButton = (props) => {
